new_tg_bot/requirements.txt
aiofiles==23.2.1
aiogram==3.17.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.4.1
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.10.0
urllib3==2.3.0
yarl==1.9.4
